% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-pop-etho.R
\name{boot_ci}
\alias{boot_ci}
\title{Bootstrap confidence interval}
\usage{
boot_ci(y, r = 5000, ci = 0.95)
}
\arguments{
\item{y}{Numeric vector}

\item{r}{Number of replicates to draw.}

\item{ci}{Confidence interval to draw from the empirical distribution.}
}
\description{
Compute the mean of a variable, and the quantiles after bootstrap resampling.
}
\seealso{
\code{boot_ci} is intended to be used as the \code{method} argument of \link{stat_pop_etho}.
Other functions, such as  \link[ggplot2:mean_se]{ggplot2::mean_se}, can be used to generate error bars.
}
