% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggetho_plot.R
\name{ggetho_plot}
\alias{ggetho_plot}
\title{\code{ggetho_plot} isolates the plotting functionality in ggetho}
\usage{
ggetho_plot(
  data,
  mapping,
  scale_x_FUN = NULL,
  discrete_y = FALSE,
  time_wrap = NULL,
  time_offset = 0,
  ...
)
}
\arguments{
\item{mapping}{A mapping object as defined in ggplot2}

\item{scale_x_FUN}{A scale function that can set the units of the x axis,
usually showing time i.e. seconds, hours, days, etc}

\item{discrete_y}{Whether the variable mapped to the y axis is discrete or not}
}
\description{
Make a ggetho plot
}
\details{
This function receives as input the outut produced by ggetho_preprocess and returns
a ggplot2 plot
}
