% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggetho_preprocess.R
\name{ggetho_preprocess}
\alias{ggetho_preprocess}
\title{Aggregate behavioral time series of binary variables into numerical summaries}
\usage{
ggetho_preprocess(
  data,
  mapping,
  summary_FUN = mean,
  summary_time_window = mins(30),
  time_wrap = NULL,
  time_offset = 0,
  multiplot = NULL,
  multiplot_period = hours(24)
)
}
\arguments{
\item{data}{An scored behavioral dataset as returned by an annotating function}

\item{summary_FUN}{A function describing a summary statistic of a binary variable}

\item{summary_time_window}{Duration of the windows upon which the statistic should be applied, in seconds}

\item{time_wrap}{For datasets encompassing more than 1 full circadian cycle (24 hours),
whether to average data belonging to the same ZT point or not}

\item{time_offset}{TODO}

\item{multiplot}{TODO}

\item{multiplot_period}{TODO}
}
\description{
Compute a summary statistic described by \code{summary_FUN} over
windows of duration \code{summary_time_window} seconds. If \code{summary_FUN} is the mean,
the computation is equivalent is equivalent
to calculating the fraction of time the animal spent on the binary's variable positive
during the window
}
