% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggperio.R
\name{ggperio}
\alias{ggperio}
\title{Prepare a ggplot object to represent periodogram data}
\usage{
ggperio(data, mapping = aes(x = period, y = power), ...)
}
\arguments{
\item{data}{\link{behavr} table containing the data and metadata}

\item{mapping}{default list of aesthetic mappings to use for plot}

\item{...}{additional arguments to be passed to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}}
}
\description{
TODO
}
\examples{
\dontrun{
library(zeitgebr)
metadata <- data.table(id=sprintf("toy_experiment|\%02d" , 1:40), region_id=1:40,
                       condition=c("A","B"),
                       sex=c("M","M", "F", "F"))
dt <- toy_activity_data(metadata, seed=107)
dt[, t := ifelse(xmv(condition) == "A", t, t * 1.01)]
per_dt <- periodogram(moving, dt, FUN = chi_sq_periodogram)
out <- ggperio(per_dt, aes(y=power, colour=condition))
out +  stat_pop_etho()
out <- ggperio(per_dt, aes(y=power - signif_threshold, colour=condition))
out +  stat_pop_etho()
out <- ggperio(per_dt, aes(y=power - signif_threshold, colour=condition))
out +  stat_pop_etho() + facet_wrap( ~ id, labeller = id_labeller)
}
}
