% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggspectro.R
\name{ggspectro}
\alias{ggspectro}
\title{Prepare a ggplot object to represent spectrogram data}
\usage{
ggspectro(data, mapping = aes(), summary_FUN = mean,
  summary_time_window = mins(30), time_wrap = NULL, time_offset = 0,
  ...)
}
\arguments{
\item{data}{\link[behavr:behavr]{behavr::behavr} table containing the data and metadata}

\item{mapping}{default list of aesthetic mappings to use for plot}

\item{summary_FUN}{method (function) used to summarise \code{variable} over time (typically, the mean)}

\item{summary_time_window}{width (in seconds) of the time window to compute a summary on}

\item{time_wrap}{time (in seconds) used to wrap the data (see details)}

\item{time_offset}{time offset (i.e. phase, in seconds) when using \code{time_wrap}}

\item{...}{additional arguments to be passed to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}}
}
\description{
This function summarises spectrogram data (containing spectrograms of multiple individual),
to show period on the \code{y} axis, time on the \code{x} axis and power on the \code{z} axis (\emph{e.g.} as a colour).
}
