% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-ld-annotations.R
\name{ld_annotation}
\alias{ld_annotation}
\title{Compute where the LD transitions occurred in an interval}
\usage{
ld_annotation(x, period = 1, phase = 0, l_ratio = 0.5)
}
\arguments{
\item{x}{Vector of length 2 containing the min and max timepoint}

\item{period}{Length of one cycle in the same units as x}

\item{phase}{Stating position within the period}

\item{l_ratio}{Numeric from 0 to 1 stating the fraction of the cycle
that corresponds to the L phase}
}
\description{
Given a range represented by its min and max, a period (cycle) length
phase (position in the cycle) and ratio of the L phase compared to the full cycle
compute the timepoints where the L->D and D->L transitions occur
}
